pyproject.toml
[project]
name = "cashflow-tracker"
version = "0.1.0"
description = ""
authors = [
    {name = "github-main-user",email = "kolya.skripn@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "django-filter (>=25.1,<26.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
flake8 = "^7.3.0"
pytest-django = "^4.11.1"
coverage = "^7.9.2"
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py313']
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 88

